image-equalize/pyproject.toml
[tool.poetry]
name = "equalizacao-histograma"
version = "1.0.0"
description = "Equaliza o histograma de uma determinada imagem de entrada."
authors = ["Olavo Nascimento <olavo.nascimento4@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.2"
opencv-python = "^4.6.0"
Pillow = "^9.2.0"
matplotlib = "^3.5.3"

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}
pylint = "^2.14.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
